Upgrade to hibernate 5.2.10.Final - first step

+++ databinder-models-hib/src/main/java/net/databinder/models/hib/HibernateProvider.java
@@ -200,7 +200,7 @@ public class HibernateProvider<T> extends PropertyDataProvider<T> {
 		if (criteriaBuilder != null)
 			criteriaBuilder.buildUnordered(crit);
 		crit.setProjection(Projections.rowCount());
-		Integer size = (Integer) crit.uniqueResult();
+		Long size = (Long) crit.uniqueResult();
 		return size == null ? 0 : size;
 	}
